skip git clean on z

+++ lib/manageiq/rpm_build/generate_core.rb
@@ -61,7 +61,7 @@ module ManageIQ
           shell_cmd("yarn install")
           shell_cmd("yarn run available-languages")
           shell_cmd("yarn run build")
-          shell_cmd("git clean -xdf")  # cleanup temp files
+          shell_cmd("git clean -xdf") if !RUBY_PLATFORM.include?("s390x") 
         end
       end
